+++ src/components/__tests__/Footer.test.tsx
@@ -22,7 +22,7 @@ jest.mock("gatsby", () => ({
   })),
 }));
 
-import Footer from "../_Footer";
+import Footer from "../Footer";
 describe("Footer", () => {
   it("renders correctly", () => {
     const tree = renderer.create(<Footer />).toJSON();
